+++ spec/dummy/client/entries/app-kitten.js
@@ -13,7 +13,7 @@ import { ButtonIcon } from 'kitten/components/buttons/button-icon'
 import { ButtonImage } from 'kitten/components/buttons/button-image'
 import ButtonImageWithTextAndBadge
   from 'kitten/components/buttons/button-image-with-text-and-badge'
-import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
+import { ButtonQuestionMarkIcon } from 'kitten/components/buttons/button-question-mark-icon'
 import {
   FacebookButtonIcon,
   LinkedinButtonIcon,
@@ -117,22 +117,29 @@ import { KarlGithubRibbon } from 'kitten/karl/headers/karl-github-ribbon'
 import { KarlMenuHeader } from 'kitten/karl/headers/karl-header'
 
 // Icons
+import { ArrowIcon } from 'kitten/components/icons/arrow-icon'
 import { CbIcon } from 'kitten/components/icons/cb-icon'
+import { CheckedCircleIcon } from 'kitten/components/icons/checked-circle-icon'
+import { CheckedIcon } from 'kitten/components/icons/checked-icon'
+import { ClockCircleIcon } from 'kitten/components/icons/clock-circle-icon'
 import { ClockIcon } from 'kitten/components/icons/clock-icon'
+import { CrossCircleIcon } from 'kitten/components/icons/cross-circle-icon'
+import { CrossIcon } from 'kitten/components/icons/cross-icon'
 import { FacebookIcon } from 'kitten/components/icons/facebook-icon'
 import GrabberIcon from 'kitten/components/icons/grabber-icon'
 import { HeartIcon } from 'kitten/components/icons/heart-icon'
-import { InfoIcon } from 'kitten/components/icons/info-icon'
-import { InformationIcon } from 'kitten/components/icons/information-icon'
-import { KissKissBankBankLogoIcon }
-  from 'kitten/components/icons/kisskissbankbanklogo-icon'
+import { InstrumentTagIcon } from 'kitten/components/icons/instrument-tag-icon'
 import { LinkedinIcon } from 'kitten/components/icons/linkedin-icon'
 import { LocationIcon } from 'kitten/components/icons/location-icon'
+import { LockIcon } from 'kitten/components/icons/lock-icon'
 import { MasterCardIcon } from 'kitten/components/icons/mastercard-icon'
 import { MenuIcon } from 'kitten/components/icons/menu-icon'
-import { RightArrowIcon } from 'kitten/components/icons/right-arrow-icon'
+import { QuestionMarkIcon } from 'kitten/components/icons/question-mark-icon'
+import { SearchIcon } from 'kitten/components/icons/search-icon'
+import { TypologyTagIcon } from 'kitten/components/icons/typology-tag-icon'
 import { TwitterIcon } from 'kitten/components/icons/twitter-icon'
 import { VisaIcon } from 'kitten/components/icons/visa-icon'
+import { WarningCircleIcon } from 'kitten/components/icons/warning-circle-icon'
 
 // Illustrations
 import { LinkBoxIllustration } from 'kitten/components/illustrations/link-box-illustration'
@@ -224,7 +231,7 @@ import {
 import { Stepper } from 'kitten/components/steppers/stepper'
 
 // Tooltips
-import { InfoWithTooltip } from 'kitten/components/tooltips/info-with-tooltip'
+import { QuestionMarkWithTooltip } from 'kitten/components/tooltips/question-mark-with-tooltip'
 import { StaticTooltip } from 'kitten/components/tooltips/static-tooltip'
 import Tooltip from 'kitten/components/tooltips/tooltip'
 import { TooltipNew } from 'kitten/components/tooltips/tooltip-new'
@@ -260,7 +267,7 @@ ReactOnRails.register({
   ButtonIcon,
   ButtonImage,
   ButtonImageWithTextAndBadge,
-  ButtonTooltipIcon,
+  ButtonQuestionMarkIcon,
   FacebookButtonIcon,
   LinkedinButtonIcon,
   TagButton,
@@ -346,21 +353,29 @@ ReactOnRails.register({
   KarlUnloggedHeaderPlatformSwitch,
 
   // Icons
+  ArrowIcon,
   CbIcon,
+  CheckedCircleIcon,
+  CheckedIcon,
+  ClockCircleIcon,
   ClockIcon,
+  CrossCircleIcon,
+  CrossIcon,
   FacebookIcon,
   GrabberIcon,
   HeartIcon,
-  InfoIcon,
-  InformationIcon,
-  KissKissBankBankLogoIcon,
+  InstrumentTagIcon,
   LinkedinIcon,
   LocationIcon,
+  LockIcon,
   MasterCardIcon,
   MenuIcon,
-  RightArrowIcon,
+  QuestionMarkIcon,
+  SearchIcon,
+  TypologyTagIcon,
   TwitterIcon,
   VisaIcon,
+  WarningCircleIcon,
 
   // Illustrations
   LinkBoxIllustration,
@@ -438,7 +453,7 @@ ReactOnRails.register({
   Stepper,
 
   // Tooltips
-  InfoWithTooltip,
+  QuestionMarkWithTooltip,
   StaticTooltip,
   Tooltip,
   TooltipNew,